Revert "Merge branch 'libusb:master' into v1.0.27"
This reverts commit c173be22611e5fe8c29148da4b827c4601130be0, reversing
changes made to 286fca62ee7eb435aba178d04fa2a852fd078508.

+++ libusb/os/threads_windows.h
@@ -105,9 +105,9 @@ static inline void usbi_tls_key_delete(usbi_tls_key_t key)
 	WINAPI_CHECK(TlsFree(key));
 }
 
-static inline unsigned long usbi_get_tid(void)
+static inline unsigned int usbi_get_tid(void)
 {
-	return (unsigned long)GetCurrentThreadId();
+	return (unsigned int)GetCurrentThreadId();
 }
 
 #endif /* LIBUSB_THREADS_WINDOWS_H */
